[Fix] 패키지 리팩터링
