--- src/main/java/com/ahnhaetdaeyo/finalbe/FinalBeApplication.java
@@ -1,15 +1,0 @@
-package com.ahnhaetdaeyo.finalbe;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@SpringBootApplication
-public class FinalBeApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(FinalBeApplication.class, args);
-    }
-
-}
