--- src/main/java/com/ahnhaetdaeyo/finalbe/dto/request/ChallengeRequestDto.java
@@ -1,16 +1,0 @@
-package com.ahnhaetdaeyo.finalbe.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class ChallengeRequestDto {
-    private String title;
-    private String leaderName;
-    private String crewName;
-    private String startDate;
-
-}
