--- src/main/java/com/ahnhaetdaeyo/finalbe/exception/SuccessMsg.java
@@ -1,4 +1,0 @@
-package com.ahnhaetdaeyo.finalbe.exception;
-
-public class SuccessMsg {
-}
