--- src/main/java/com/ahnhaetdaeyo/finalbe/security/configuration/WebSecurityConfig.java
@@ -1,62 +1,0 @@
-package com.ahnhaetdaeyo.finalbe.security.configuration;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-@EnableWebSecurity
-@Configuration
-@RequiredArgsConstructor
-public class WebSecurityConfig {
-
-
-    @Bean
-    // h2-console 사용에 대한 허용 (CSRF, FrameOptions 무시)
-    public WebSecurityCustomizer webSecurityCustomizer() {
-        return (web) -> web.ignoring()
-                .antMatchers("/h2-console/**");
-    }
-
-    @Bean
-    SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.cors().configurationSource(corsConfigurationSource());
-
-        http.csrf().disable();
-
-        http.authorizeRequests((authz)->authz
-                .antMatchers("/signup/**").permitAll()
-                .anyRequest().permitAll());
-
-        return http.build();
-    }
-
-    @Bean
-    CorsConfigurationSource corsConfigurationSource() {
-        CorsConfiguration configuration = new CorsConfiguration();
-
-        //허용할 url 설정
-        configuration.addAllowedOrigin("http://localhost:3000");
-        //허용할 헤더 설정
-        configuration.addAllowedHeader("*");
-        //허용할 http method
-        configuration.addAllowedMethod("*");
-        //사용자 자격 증명이 지원되는지 여부
-        configuration.setAllowCredentials(true);
-
-        // 클라이언트가 응답에 접근할 수 있는 헤더 설정
-        configuration.addExposedHeader("Authorization");
-        configuration.addExposedHeader("Refresh-Token");
-
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-        return source;
-
-    }
-}
