+++ src/main/java/com/havit/finalbe/dto/request/ChallengeRequestDto.java
@@ -1,0 +1,16 @@
+package com.havit.finalbe.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class ChallengeRequestDto {
+    private String title;
+    private String leaderName;
+    private String crewName;
+    private String startDate;
+
+}
