+++ src/main/java/com/havit/finalbe/entity/Challenge.java
@@ -1,0 +1,39 @@
+package com.havit.finalbe.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+public class Challenge extends Timestamped{
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long challengeId;
+
+    @Column(nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private String nickname;
+
+    @Column(nullable = false)
+    private String leaderName;
+
+    @Column(nullable = false)
+    private String crewName;
+
+    @Column(nullable = false)
+    private String startDate;
+
+    @Column(nullable = false)
+    private String imgUrl;
+
+}
