+++ src/main/java/com/havit/finalbe/exception/SuccessMsg.java
@@ -1,0 +1,4 @@
+package com.havit.finalbe.exception;
+
+public class SuccessMsg {
+}
