+++ src/main/java/com/havit/finalbe/security/userDetail/UserDetailService.java
@@ -1,0 +1,4 @@
+package com.havit.finalbe.security.userDetail;
+
+public class UserDetailService {
+}
