--- src/test/java/com/ahnhaetdaeyo/finalbe/FinalBeApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ahnhaetdaeyo.finalbe;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FinalBeApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
