+++ src/test/java/com/havit/finalbe/FinalBeApplicationTests.java
@@ -1,0 +1,13 @@
+package com.havit.finalbe;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class FinalBeApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
